requirements.txt
streamlit
langchain
langchain-core
langchain-openai
openai
python-dotenv
langchain_community
langserve
fastapi
unicorn
sse_starlette